refactor: extract ffmpeg input resolver from adapter
